Finished close settlementWindow
